feat: connect server to deployed database

+++ server/src/main/java/com/train/tinder/repository/TrainRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.train.tinder.model.Train;
 
-public interface TrainRepository extends JpaRepository<Train, Long>{
-    
+public interface TrainRepository extends JpaRepository<Train, Integer> {
+
 }
